REFACTOR: core/pkg/jwt -> core/internal/pkg/jwt
