+++ core/internal/pkg/jwt/jwt.go
@@ -1,7 +1,6 @@
 package jwt
 
 import (
-	// TODO move to local
 	cons "core/internal/pkg/constants"
 	rsc "core/internal/pkg/resource"
 	"errors"
